backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pandas
openpyxl
numpy
matplotlib
firebase_admin
openpyxl
google-api-python-client
python-multipart